+++ components/site/tests/site.rs
@@ -666,11 +666,7 @@ fn can_build_with_extra_syntaxes() {
 
     assert!(&public.exists());
     assert!(file_exists!(public, "posts/extra-syntax/index.html"));
-    assert!(file_contains!(
-        public,
-        "posts/extra-syntax/index.html",
-        r#"<span style="color:"#
-    ));
+    assert!(file_contains!(public, "posts/extra-syntax/index.html", r#"<span style="color:"#));
 }
 
 #[test]