+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Allow specifying default language in filenames
 - Render emoji in Markdown content if the `render_emoji` option is enabled
 - Enable YouTube privacy mode for the YouTube shortcode
-- Add language as class to the `<code>` block
+- Add language as class to the `<code>` block and as `data-lang`
 - Add bibtex to `load_data`
 - Add a `[markdown]` section to `config.toml` to configure rendering
 - Add `highlight_code` and `highlight_theme` to a `[markdown]` section in `config.toml`